[refactor] MVC 구조에 맞춰 리팩토링
