+++ src/main/java/com/example/aniwhere/controller/dto/AnimeQuarterDTO.java
@@ -1,4 +1,4 @@
-package com.example.aniwhere.domain.anime.dto;
+package com.example.aniwhere.controller.dto;
 
 
 import lombok.Builder;
