+++ src/main/java/com/example/aniwhere/repository/VoiceactorRepository.java
@@ -1,9 +1,7 @@
-package com.example.aniwhere.application.voiceactor.repository;
+package com.example.aniwhere.repository;
 
 import com.example.aniwhere.domain.voiceactor.VoiceActor;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 
 public interface VoiceactorRepository extends JpaRepository<VoiceActor, Long> {
 }
